requirements.txt
transformers
torch
asyncio
openai
httpx
python-dotenv
aiogram